Read emsg when reading fragmented header
