+++ src/cmaf/mod.rs
@@ -1,2 +1,3 @@
 pub mod cmaf_chunk_writer;
 pub mod cmaf_header_writer;
+pub mod emsg;
